LibMedia: Move FFmpegIOContext into it's own file
This allows it to be reused for video.
